reorganise glow util
