+++ src/glow_util/FpsCamera.cpp
@@ -1,5 +1,5 @@
 // Copyright 2017 Emanuele Palazzolo (emanuele.palazzolo@uni-bonn.de)
-#include "glow_util/FpsCamera.h"
+#include "glow/util/FpsCamera.h"
 
 #include <glow/glutil.h>
