+++ src/glow_util/GlCamera.cpp
@@ -1,4 +1,4 @@
-#include "glow_util/GlCamera.h"
+#include "glow/util/GlCamera.h"
 
 #include "glow/glutil.h"
