+++ src/glow_util/X11OffscreenContext.cpp
@@ -1,4 +1,4 @@
-#include "glow_util/X11OffscreenContext.h"
+#include "glow/util/X11OffscreenContext.h"
 
 #include <iostream>
 #include <stdexcept>
